added model api.
This change makes the entire model api self-contained inside the models.js file, with the exception of the 'defineModel' function in the root of the library. That's fine; it means that using models are entirely an optional thing, but are just as easy to do as low-level database access. Provided the unit tests for this work out well, this should be ready to go shortly.