added support for automatic model syncing.
Now, the database object emits a 'sync' event whenever the `sync` function is called. Models now listen for this sync function, and update themselves, if they are not dirty. This ensures that model objects will be as up to date as possible.

If you want to force a model instance to sync, you can call it's `sync()`, passing in the boolean `true`. This will force a sync, regardless of if the mode is dirty.